Completed with backed of cart.php

+++ _headers/alterCartData.php
@@ -7,14 +7,15 @@
 
     global $con;
 
-    if( isset( $_POST['delete'] ) && isset( $_SESSION['ID'] ) && $_POST['delete'] != "" ){
+    if( isset( $_POST['delete'] ) && isset( $_SESSION['ID'] ) && $_POST['delete'] != "" && !( isset( $_POST['updateQuantity'] ) ) && !( isset( $_POST['atc'] ) ) ){
         $lid = mysqli_escape_string( $con, $_SESSION['ID'] );
         $delete = mysqli_escape_string( $con, $_POST['delete'] );
         $sql = "DELETE FROM cart WHERE CRT_LOGIN_ID = '" . $lid . "' AND CRT_PRD_ID = '" . $delete . "'; ";
         mysqli_query( $con, $sql );
+        die();
     }
 
-    if( isset( $_POST['token'] )  && isset( $_SESSION['ID'] ) && isset( $_POST['updateQuantity'] ) && $_POST['token'] != "" && $_POST['updateQuantity'] != "" ){
+    if( isset( $_POST['token'] )  && isset( $_SESSION['ID'] ) && isset( $_POST['updateQuantity'] ) && $_POST['token'] != "" && $_POST['updateQuantity'] != "" && !( isset( $_POST['delete'] ) )  && !( isset( $_POST['atc'] ) ) ){
 
         if( $_POST['updateQuantity'] >= 1 && $_POST['updateQuantity'] < 5 ){
             $pid = mysqli_escape_string( $con, $_POST['token'] );
@@ -26,8 +27,28 @@
             $sql .= "WHERE CRT_PRD_ID = '" . $pid . "' AND ";
             $sql .= "CRT_LOGIN_ID = '" . $lid . "'; ";
             mysqli_query( $con, $sql );
-            echo " ";
+            die(" ");
         } else {
-            echo "Quantity cannot be 0 and should be less than 5";
+            die("Quantity cannot be 0 and should be less than 5");
+        }
+    }
+
+    if(  isset( $_POST['atc'] ) && isset( $_SESSION['ID'] ) && $_POST['atc'] != "" && !( isset( $_POST['updateQuantity'] ) ) && !( isset( $_POST['delete'] ) ) ){
+
+        $pid = mysqli_escape_string( $con, $_POST['atc'] );
+        $lid = mysqli_escape_string( $con, $_SESSION['ID'] );
+
+        $sql = "SELECT * FROM cart WHERE CRT_LOGIN_ID = '" . $lid . "' AND CRT_PRD_ID = '" . $pid . "'; ";
+
+        if( mysqli_num_rows( mysqli_query( $con, $sql ) ) > 0 ){
+            die("Product exists in your cart!");
+        } else {
+            $sql = "INSERT INTO cart VALUES ";
+            $sql .= "( '" . $lid . "', ";
+            $sql .= "'".$pid."', 1); ";
+
+            mysqli_query( $con, $sql );
+
+            die("Added to cart!");
         }
     }
